refactor(rag): move context_service utilities into retrieval/utils.py
All three functions in context_service.py were retrieval utilities with no
reason to live at the top level of app/rag/:

  - _lookup_drive_file    → retrieval/utils.py (used by agent.py)
  - format_chunks_as_context → retrieval/utils.py (used by agent.py)
  - _validate_files_ready → retrieval/utils.py (used by app/lib/rag.py)

The lazy import that was inside get_context_via_agent (to avoid a circular
import via context_service) becomes a module-level import in agent.py now
that both files live in the same package — no cycle exists within retrieval/.
The lazy DriveFile import moves into retrieval/utils.py to preserve the
existing cycle protection (app.models.drive → app.lib.rag → retrieval/utils
→ app.models.drive at load time).

context_service.py is deleted.
