+++ app/lib/rag.py
@@ -10,7 +10,6 @@ from langchain_core.language_models import BaseChatModel
 
 from app.lib.db import session_scope
 from app.memory_profiler import profile_memory
-from app.rag.context_service import _validate_files_ready
 from app.rag.enums import RagStatus  # noqa: F401 — re-exported in __all__
 from app.rag.exceptions import (  # noqa: F401 — re-exported in __all__
     DriveFileNotFoundError,
@@ -22,6 +21,7 @@ from app.rag.exceptions import (  # noqa: F401 — re-exported in __all__
 from app.rag.ingestion.chunking import DocumentChunker
 from app.rag.ingestion.extraction import check_extraction_eligibility, extract_to_markdown
 from app.rag.retrieval import get_context_via_agent_with_isolated_session
+from app.rag.retrieval.utils import _validate_files_ready
 from app.rag.store import ChunkStoreService, store_and_link_chunks
 from app.rag.types import IngestionResult, RetrievedChunk  # noqa: F401 — re-exported in __all__
